[+] `LedgerLoginContainer` refactor
