+++ src/UI/SignTransactionsModals/SignWithDeviceModal/components/SignStepBody.tsx
@@ -7,10 +7,10 @@ import { useGetTokenDetails } from 'hooks/transactions/useGetTokenDetails';
 import { ActiveLedgerTransactionType, MultiSignTransactionType } from 'types';
 import { NftEnumType } from 'types/tokens.types';
 import { TransactionData } from 'UI/TransactionData';
-import { getIdentifierType } from 'utils';
 import { getEgldLabel } from 'utils/network/getEgldLabel';
 import { formatAmount } from 'utils/operations/formatAmount';
 import { isTokenTransfer } from 'utils/transactions/isTokenTransfer';
+import { getIdentifierType } from 'utils/validation/getIdentifierType';
 
 import { useSignStepsClasses } from '../hooks';
 import { ConfirmAmount } from './components/ConfirmAmount';
