+++ src/UI/ledger/LedgerLoginContainer/AddressTable/index.ts
@@ -1,0 +1,1 @@
+export * from './AddressTable';
