+++ src/UI/ledger/LedgerLoginContainer/ConfirmAddress/index.ts
@@ -1,0 +1,1 @@
+export * from './ConfirmAddress';
