+++ src/UI/ledger/LedgerLoginContainer/LedgerConnect/LedgerConnect.tsx
@@ -5,7 +5,7 @@ import LedgerLogo from 'assets/icons/ledger-nano.svg';
 
 import globalStyles from 'assets/sass/main.scss';
 import { DataTestIdsEnum } from 'constants/index';
-import { WithClassnameType } from '../../types';
+import { WithClassnameType } from '../../../types';
 
 import styles from './ledgerConnectStyles.scss';
