+++ src/UI/ledger/LedgerLoginContainer/LedgerScamPhishingAlert/index.ts
@@ -1,0 +1,1 @@
+export * from './LedgerScamPhishingAlert';
